backend/firstlayer/category_classifier/requirements.txt
# FirstLayer 问题分类服务 - 依赖包
# 一次性安装所有相关库

# Web 框架
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
pydantic>=2.5.0

# 环境变量
python-dotenv>=1.0.0

# ML 模型（可选，已下载）
transformers>=4.36.0
torch>=2.2.0
accelerate>=0.26.0
sentencepiece>=0.1.99

# 工具库
requests>=2.31.0
httpx>=0.26.0
